CAMEL-18962 refactor existing tests for server security enablement
